chore: refactored the unit tests for common code
